Rename simulation modules
