+++ src/sim/temporal/constant.py
