+++ src/sim/temporal/descending.py
